tests/wpt/web-platform-tests/css/css-align/parsing/justify-tracks-valid.html

<!DOCTYPE html>
<html>
<head>
<meta charset="utf-8">
<title>CSS Grid Level 3: parsing justify-tracks with valid values</title>
<link rel="help" href="https://drafts.csswg.org/css-grid-3/#tracks-alignment">
<meta name="assert" content="justify-tracks supports the full grammar 'normal | <content-distribution> | <overflow-position>? [ <content-position> | left | right ]'.">
<script src="/resources/testharness.js"></script>
<script src="/resources/testharnessreport.js"></script>
<script src="/css/support/parsing-testcommon.js"></script>
</head>
<body>
<script>
test_valid_value("justify-tracks", "normal");

// <content-distribution> = space-between | space-around | space-evenly | stretch
test_valid_value("justify-tracks", "space-between");
test_valid_value("justify-tracks", "space-around");
test_valid_value("justify-tracks", "space-evenly");
test_valid_value("justify-tracks", "stretch");

// <overflow-position>? [ <content-position> | left | right ]
// <overflow-position> = unsafe | safe
// <content-position> = center | start | end | flex-start | flex-end
test_valid_value("justify-tracks", "center");
test_valid_value("justify-tracks", "start");
test_valid_value("justify-tracks", "end");
test_valid_value("justify-tracks", "flex-start");
test_valid_value("justify-tracks", "flex-end");
test_valid_value("justify-tracks", "unsafe end");
test_valid_value("justify-tracks", "safe flex-start");
test_valid_value("justify-tracks", "left");
test_valid_value("justify-tracks", "unsafe right");

test_valid_value("justify-tracks", "flex-start, last baseline");
test_valid_value("justify-tracks", "normal, normal", "normal");
test_valid_value("justify-tracks", "start, flex-end, flex-end, flex-end, flex-end", "start, flex-end");

</script>
</body>
</html>
